fixes and modularity
