+++ src/controllers/blog.py
@@ -1,10 +1,10 @@
 from src.controllers.post import Post
-from src.common.sql_query import Sql
+from src.config.sql_query import Sql
 from src.models.database import Database
 from src.controllers.comment import Comment
 from src.loggers.general_logger import GeneralLogger
-from src.common import filepaths
-from src.common import prompts
+from src.config import filepaths
+from src.config import prompts
 
 
 class Blog(Post):
